requirements.txt
aiofiles==23.1.0
aiogram==3.1.1
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.6.2
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2023.7.22
charset-normalizer==3.0.1
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.2.0
cycler==0.12.1
DAWG-Python==0.7.2
debugpy==1.8.1
decorator==5.1.1
docopt==0.6.2
et-xmlfile==1.1.0
exceptiongroup==1.2.1
executing==2.0.1
fonttools==4.49.0
frozenlist==1.4.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.16.3
httpx==0.23.3
hyperframe==6.0.1
idna==3.4
ipykernel==6.29.4
ipython==8.23.0
jedi==0.19.1
joblib==1.3.2
jupyter_client==8.6.1
jupyter_core==5.7.2
kiwisolver==1.4.5
magic-filter==1.0.12
matplotlib==3.8.3
matplotlib-inline==0.1.7
multidict==6.0.4
nest-asyncio==1.6.0
nltk==3.8.1
numpy==1.26.0
openpyxl==3.1.3
packaging==24.0
pandas==2.2.1
parso==0.8.4
patsy==0.5.6
pillow==10.2.0
platformdirs==4.2.1
prompt-toolkit==3.0.43
psutil==5.9.8
pure-eval==0.2.2
pydantic==2.3.0
pydantic_core==2.6.3
Pygments==2.17.2
pymorphy2==0.9.1
pymorphy2-dicts-ru==2.4.417127.4579844
pyparsing==3.1.2
pyTelegramBotAPI==4.10.0
python-dateutil==2.9.0.post0
python-telegram-bot==20.1
pytz==2024.1
pywin32==306
pyzmq==26.0.2
regex==2024.5.15
requests==2.28.2
rfc3986==1.5.0
scikit-learn==1.4.1.post1
scipy==1.12.0
seaborn==0.13.2
six==1.16.0
sniffio==1.3.0
socksio==1.0.0
soupsieve==2.5
stack-data==0.6.3
statsmodels==0.14.2
threadpoolctl==3.3.0
tornado==6.4
tqdm==4.66.4
traitlets==5.14.3
typing_extensions==4.7.1
tzdata==2024.1
urllib3==1.26.14
wcwidth==0.2.13
wordcloud==1.9.3
yarl==1.9.2
